A function to get ordinal number with test
