+++ Sprint-3/revise/implement/get-ordinal-number.test.js
@@ -15,27 +15,8 @@
    // get-ordinal-number.test.js (or your file name)
 
 // The function implementation
-function getOrdinalNumber(number) {
-    const remainder10 = number % 10;
-    const remainder100 = number % 100;
 
-    if (remainder100 >= 11 && remainder100 <= 13) {
-        return `${number}th`;
-    }
-
-    switch (remainder10) {
-        case 1:
-            return `${number}st`;
-        case 2:
-            return `${number}nd`;
-        case 3:
-            return `${number}rd`;
-        default:
-            return `${number}th`;
-    }
-}
-
-
+const getOrdinalNumber = require('./get-ordinal-number');
 describe('getOrdinalNumber', () => {
     it('should return "1st" for 1', () => {
         expect(getOrdinalNumber(1)).toBe('1st');
